+++ healvis/observatory.py
@@ -6,7 +6,6 @@ from __future__ import absolute_import, division, print_function
 
 import numpy as np
 import multiprocessing as mp
-from numba import jit
 import sys
 import resource
 import warnings
@@ -27,7 +26,6 @@ from .cosmology import c_ms
 # -----------------------
 
 
-@jit
 def make_fringe(az, za, freq, enu):
     """
     az, za = Azimuth, zenith angle, radians