+++ healvis/sky_model.py
@@ -6,7 +6,6 @@ from __future__ import absolute_import, division, print_function
 
 import numpy as np
 import os
-import six
 import warnings
 import healpy as hp
 with warnings.catch_warnings():  # noqa
@@ -35,12 +34,12 @@ class SkyModel(object):
     """
     SkyModel class
     """
-    valid_params = ['Npix', 'Nside', 'Nskies', 'Nfreqs', 'indices', 'Z_array', 'ref_chan',
-                    'pspec_amp', 'freqs', 'data', 'history']
+    valid_params = ['Npix', 'Nside', 'Nskies', 'Nfreqs', 'indices', 'Z_array',
+                    'ref_chan', 'ref_freq', 'pspec_amp', 'freqs', 'data', 'history']
     _updated = []
     # keys give HDF5 datasets, value give their dtype
     dsets = {'data': np.float64, 'indices': np.int32, 'freqs': np.float64,
-             'history': h5py.special_dtype(vlen=six.text_type)}
+             'history': h5py.special_dtype(vlen=str)}
 
     def _defaults(self):
         """
@@ -142,6 +141,7 @@ class SkyModel(object):
 
         self.data = flat_spectrum_noise_shell(sigma, self.freqs, self.Nside, self.Nskies,
                                               ref_chan=self.ref_chan, shared_memory=shared_memory)
+        self.ref_freq = self.freqs[self.ref_chan]
         self.pspec_amp = sigma
         self._update()
 
@@ -171,14 +171,17 @@ class SkyModel(object):
 
         print('...reading {}'.format(filename))
         with h5py.File(filename, 'r') as infile:
-
             if do_not_overwrite_freqs:
                 sky_freqs_full = infile['freqs'][()]
-                freq_chans = np.where(np.in1d(sky_freqs_full, self.freqs))[0]
+                # Find nearest frequency in sky_freqs for each self.freqs.
+                freq_chans = []
+                for fr in self.freqs:
+                    freq_chans.append(np.argmin(np.abs(fr - sky_freqs_full)))
+                freq_chans = np.sort(freq_chans)
                 sky_freqs_part = sky_freqs_full[freq_chans]
                 if self.freqs is not None and not np.allclose(self.freqs, sky_freqs_part):
                     raise ValueError("Currently set frequencies do not match any subset of file's frequencies.")
-                Nfreqs_load = freq_chans.size
+                Nfreqs_load = len(freq_chans)
 
             # load lightweight attributes
             for k in infile.attrs:
@@ -193,21 +196,31 @@ class SkyModel(object):
                             if Nfreqs_load is not None:
                                 s[-1] = Nfreqs_load
                             s = tuple(s)
+                            if len(s) < 3:
+                                s = (1,) + s
                             self.data = mparray(s, dtype=np.float)
-                            self.data[()] = infile[k][:, :, freq_chans]   # Transfer data from infile.
+                            self.data[()] = infile[k][..., freq_chans]   # Transfer data from infile.
                         else:
                             # load the data via slice
                             setattr(self, k, infile[k][:, :, freq_chans])
                     elif k == 'freqs':
                         setattr(self, k, infile[k][:][freq_chans])
                     elif k == 'history':
-                        setattr(self, k, infile[k].value)
+                        setattr(self, k, infile[k][()])
                     else:
                         setattr(self, k, infile[k][:])
+
+            if self.ref_freq is None:
+                self.ref_freq = infile['freqs'][:][self.ref_chan]
 
             # make sure Nfreq agrees
             self.Nfreqs = len(self.freqs)
 
+        if self.Nside is None:
+            try:
+                self.Nside = hp.npix2nside(self.data.shape[1])
+            except(ValueError):
+                raise ValueError("Data array is not a full HEALPix map, and Nside not provided.")
         self._update()
 
     def write_hdf5(self, filename, clobber=False):